requirements.txt
# Spotify Playlist Creator App
streamlit>=1.30.0
spotipy>=2.23.0
python-dotenv>=1.0.0
